+++ CHANGELOG.md
@@ -218,5 +218,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - the package configuration file finds the libraries the module links, so that
   a project using the installed module needs nothing more than find_package()
 
-[Unreleased]: https://gitlab.com/smspp/svmblock/-/compare/0.1.0...develop
+[Unreleased]: https://gitlab.com/smspp/svmblock/-/compare/0.2.0...develop
+[0.2.0]: https://gitlab.com/smspp/svmblock/-/compare/0.1.0...0.2.0
 [0.1.0]: https://gitlab.com/smspp/svmblock/-/tags/0.1.0